+++ src/http/gui/app/routes/home/index.js
@@ -1,9 +1,0 @@
-import PageBase from './_page-base';
-
-export default PageBase.extend({
-  name: 'index',
-
-  beforeModel() {
-    this.transitionTo('onezone')
-  }
-});