+++ src/logic/token_logic.erl
@@ -13,7 +13,7 @@
 -author("Konrad Zemek").
 
 -include("registered_names.hrl").
--include("handlers/rest_handler.hrl").
+-include("http/handlers/rest_handler.hrl").
 -include("datastore/oz_datastore_models_def.hrl").
 -include_lib("ctool/include/logging.hrl").