+++ src/subscriptions/translator.erl
@@ -16,7 +16,7 @@
 -include("datastore/oz_datastore_models_def.hrl").
 -include_lib("ctool/include/logging.hrl").
 
--export([get_msg/3, get_ignore_msg/1]).
+-export([get_ignore_msg/1, as_msg/3]).
 
 %%%-------------------------------------------------------------------
 %%% @doc
@@ -24,19 +24,47 @@
 %%% Those structures are serializable to json.
 %%% @end
 %%%-------------------------------------------------------------------
--spec get_ignore_msg(Seq :: pos_integer())
-        -> term().
+-spec as_msg(Seq :: subscriptions:seq(), Doc :: datastore:document(),
+    Ignore :: boolean()) -> term().
+as_msg(Seq, Doc = #document{value = Value}, false) ->
+    get_msg(Seq, Doc, element(1, Value));
+as_msg(-1, Doc = #document{value = Value}, _) ->
+    get_msg(-1, Doc, element(1, Value));
+as_msg(Seq, Doc = #document{value = #onedata_user{}}, true) ->
+    #document{value = Value = #onedata_user{name = Name}, key = ID} = Doc,
+    Model = element(1, Value),
+    [{seq, Seq}, revs_prop(Doc), {id, ID}, {message_model(Model), [
+        {name, Name},
+        {space_ids, []},
+        {group_ids, []},
+        {default_space, undefined},
+        {public_only, true}
+    ]}];
+as_msg(Seq, _Doc, _Ignore) ->
+    get_ignore_msg(Seq).
 
+%%%-------------------------------------------------------------------
+%%% @doc
+%%% Translates documents to structures required by the provider.
+%%% Those structures are serializable to json and contain "ignore" commands.
+%%% @end
+%%%-------------------------------------------------------------------
+-spec get_ignore_msg(Seq :: subscriptions:seq()) -> term().
 get_ignore_msg(Seq) ->
     [{seq, Seq}, {ignore, true}].
 
+%%%===================================================================
+%%% Internal functions
+%%%===================================================================
+
 %%%-------------------------------------------------------------------
 %%% @doc
+%%% @private
 %%% Translates documents to structures required by the provider.
-%%% Those structures are serializable to json.
+%%% Those structures are serializable to json and provide details from documents.
 %%% @end
 %%%-------------------------------------------------------------------
--spec get_msg(Seq :: pos_integer(), Doc :: datastore:document(),
+-spec get_msg(Seq :: subscriptions:seq(), Doc :: datastore:document(),
     Model :: subscriptions:model()) -> term().
 
 get_msg(Seq, Doc = #document{deleted = true, key = ID}, Model) ->
@@ -44,12 +72,11 @@ get_msg(Seq, Doc = #document{deleted = true, key = ID}, Model) ->
 get_msg(Seq, Doc, space = Model) ->
     #document{value = Value, key = ID} = Doc,
     #space{name = Name, users = Users, groups = Groups,
-        providers = Providers, size = Sizes} = Value,
+        providers_supports = Supports} = Value,
     [{seq, Seq}, revs_prop(Doc), {id, ID}, {message_model(Model), [
         {id, ID},
         {name, Name},
-        {size, Sizes},
-        {providers, Providers},
+        {providers_supports, Supports},
         {users, Users},
         {groups, Groups}
     ]}];
@@ -63,18 +90,18 @@ get_msg(Seq, Doc, user_group = Model) ->
     ]}];
 get_msg(Seq, Doc, onedata_user = Model) ->
     #document{value = Value, key = ID} = Doc,
-    #onedata_user{name = Name, spaces = Spaces, groups = Groups} = Value,
+    #onedata_user{name = Name, space_names = SpaceNames, groups = Groups,
+        default_space = DefaultSpace} = Value,
     [{seq, Seq}, revs_prop(Doc), {id, ID}, {message_model(Model), [
         {name, Name},
-        {space_ids, Spaces},
-        {group_ids, Groups}
+        {space_names, maps:to_list(SpaceNames)},
+        {group_ids, Groups},
+        {default_space, DefaultSpace},
+        {public_only, false}
     ]}];
 get_msg(_Seq, _Doc, _Model) ->
+    ?warning("Requesting message for unexpected model ~p", [_Model]),
     [].
-
-%%%===================================================================
-%%% Internal functions
-%%%===================================================================
 
 -spec revs_prop(Doc :: datastore:document()) -> term().
 revs_prop(#document{rev = Revs}) when is_tuple(Revs) ->