+++ src/utils/dev_utils.erl
@@ -52,14 +52,14 @@
 
 -include_lib("ctool/include/logging.hrl").
 -include("datastore/oz_datastore_models_def.hrl").
--include("handlers/rest_handler.hrl").
+-include("http/handlers/rest_handler.hrl").
 
 %% API
 -export([set_up_test_entities/3, destroy_test_entities/3]).
 -export([create_provider_with_uuid/6, create_provider_with_uuid/5]).
 -export([create_user_with_uuid/2]).
 -export([create_group_with_uuid/3]).
--export([create_space_with_uuid/3, create_space_with_uuid/5, create_space_with_provider/5]).
+-export([create_space_with_uuid/3, create_space_with_uuid/5, create_space_with_provider/4]).
 
 %%%===================================================================
 %%% API
@@ -155,6 +155,14 @@ set_up_test_entities(Users, Groups, Spaces) ->
                         space_logic:join({group, GroupID}, SpaceToken)
                     end, GroupsToAdd)
             end, Spaces),
+
+        % Give all space perms to users that have it as default space
+        lists:foreach(
+            fun({UserID, Props}) ->
+                DefaultSpace = proplists:get_value(<<"default_space">>, Props),
+                space_logic:set_privileges(
+                    DefaultSpace, {user, UserID}, privileges:space_admin())
+            end, Users),
         ok
     catch
         T:M ->
@@ -257,7 +265,7 @@ create_group_with_uuid(UserId, Name, UUID) ->
 -spec create_space_with_uuid({user | group, Id :: binary()}, Name :: binary(), UUID :: binary()) ->
     {ok, SpaceId :: binary()} | no_return().
 create_space_with_uuid(Member, Name, UUID) ->
-    create_space_with_provider(Member, Name, [], [], UUID).
+    create_space_with_provider(Member, Name, [], UUID).
 
 
 %%--------------------------------------------------------------------
@@ -266,12 +274,12 @@ create_space_with_uuid(Member, Name, UUID) ->
 %% @end
 %%--------------------------------------------------------------------
 -spec create_space_with_uuid({provider, ProviderId :: binary()}, Name :: binary(),
-    Token :: binary(), Size :: pos_integer(), UUID :: binary()) ->
+    Token :: binary(), Support :: pos_integer(), UUID :: binary()) ->
     {ok, SpaceId :: binary()}.
-create_space_with_uuid({provider, ProviderId}, Name, Token, Size, UUID) ->
+create_space_with_uuid({provider, ProviderId}, Name, Token, Support, UUID) ->
     {ok, Macaroon} = macaroon:deserialize(Token),
     {ok, Member} = token_logic:consume(Macaroon),
-    create_space_with_provider(Member, Name, [ProviderId], [{ProviderId, Size}], UUID).
+    create_space_with_provider(Member, Name, [{ProviderId, Support}], UUID).
 
 
 %%--------------------------------------------------------------------
@@ -279,28 +287,34 @@ create_space_with_uuid({provider, ProviderId}, Name, Token, Size, UUID) ->
 %% Throws exception when call to the datastore fails, or user/group doesn't exist.
 %% @end
 %%--------------------------------------------------------------------
--spec create_space_with_provider({user | group, Id :: binary()}, Name :: binary(), Providers :: [binary()],
-    Size :: [{Provider :: binary(), ProvidedSize :: pos_integer()}], UUID :: binary()) ->
+-spec create_space_with_provider({user | group, Id :: binary()}, Name :: binary(),
+    Support :: [{Provider :: binary(), ProvidedSize :: pos_integer()}], UUID :: binary()) ->
     {ok, SpaceId :: binary()}.
-create_space_with_provider({user, UserId}, Name, Providers, Size, UUID) ->
+create_space_with_provider({user, UserId}, Name, Support, UUID) ->
     {ok, UserDoc} = onedata_user:get(UserId),
     #document{value = #onedata_user{spaces = Spaces} = User} = UserDoc,
 
     Privileges = privileges:space_admin(),
-    Space = #space{name = Name, size = Size, providers = Providers, users = [{UserId, Privileges}]},
+    Space = #space{name = Name, providers_supports = Support, users = [{UserId, Privileges}]},
     {ok, SpaceId} = space:save(#document{key = UUID, value = Space}),
     UserNew = User#onedata_user{spaces = [SpaceId | Spaces]},
     onedata_user:save(UserDoc#document{value = UserNew}),
+    user_logic:set_space_name_mapping(UserId, SpaceId, Name),
 
     {ok, SpaceId};
-create_space_with_provider({group, GroupId}, Name, Providers, Size, UUID) ->
+create_space_with_provider({group, GroupId}, Name, Support, UUID) ->
     {ok, GroupDoc} = user_group:get(GroupId),
-    #document{value = #user_group{spaces = Spaces} = Group} = GroupDoc,
+    #document{value = #user_group{users = Users, spaces = Spaces} = Group} = GroupDoc,
 
     Privileges = privileges:space_admin(),
-    Space = #space{name = Name, size = Size, providers = Providers, groups = [{GroupId, Privileges}]},
+    Space = #space{name = Name, providers_supports = Support,
+        groups = [{GroupId, Privileges}]},
     {ok, SpaceId} = space:save(#document{key = UUID, value = Space}),
     GroupNew = Group#user_group{spaces = [SpaceId | Spaces]},
     user_group:save(GroupDoc#document{value = GroupNew}),
+
+    lists:foreach(fun({UserId, _}) ->
+        user_logic:set_space_name_mapping(UserId, SpaceId, Name)
+    end, Users),
 
     {ok, SpaceId}.